requirements.txt
Flask-SQLAlchemy==2.4.3
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
boto3==1.13.5
botocore==1.16.7
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.24.2
pycparser==2.19
PyMySQL==0.9.3
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.0
s3transfer==0.3.0
urllib3==1.25.3
Werkzeug==0.15.4
pytest==5.3.5

nltk==3.4.5
scikit-learn==0.23.1
xgboost==0.90
